mod: changes flow how to register to get limit. and fix calculate from transaction

+++ internal/pkg/constant/fee_constant.go
@@ -2,5 +2,5 @@ package constant
 
 const (
 	AdminFee float64 = 5000
-	Interest int     = 10
+	Interest float64 = 10.0
 )
